requirements.txt
gradio
networkx
matplotlib
pillow
rdflib
python-dotenv